fed-scaffold/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flair-federated"
version = "1.0.0"
description = "Federated Learning on FLAIR dataset with MobileNetV2"
dependencies = [
    "flwr[simulation]>=1.11.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "datasets>=2.14.0",
    "transformers>=4.30.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "numpy>=1.24.0",
    "Pillow>=9.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app.config]
num-server-rounds = 100
fraction-fit = 0.1
fraction-evaluate = 0
min-fit-clients = 20
min-evaluate-clients = 0
min-available-clients = 200

[tool.flwr.app]
publisher = "flwr"

[tool.flwr.app.components]
serverapp = "fed_scaffold.server_app:app"
clientapp = "fed_scaffold.client_app:app"

[tool.flwr.federations]
default = "local-simulation"

[tool.flwr.federations.local-simulation]
options.num-supernodes = 200
options.backend.client-resources.num-cpus = 1
options.backend.client-resources.num-gpus = 0.5
